BugID:21716117: Fix vfs test compile error for esp8266

+++ test/testcase/kernel/fs/vfs/vfs_test.c
@@ -39,7 +39,7 @@ static int32_t vfs_test_write(vfs_file_t *fp, const char *buf, uint32_t len)
     return VFS_FUNC_ATTACH;
 }
 
-static uint32_t vfs_test_lseek(vfs_file_t *fp, uint32_t off, int32_t whence)
+static uint32_t vfs_test_lseek(vfs_file_t *fp, int64_t off, int32_t whence)
 {
     return VFS_FUNC_ATTACH;
 }
